third_party/blink/web_tests/external/wpt/fullscreen/api/navigate-iframe.sub-expected.txt

This is a testharness.js-based test.
[FAIL] cross_to_same
  assert_equals: expected false but got true
Harness: the test ran to completion.
